% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_subfunction.r
\name{Abun_CreatBootstrapSample}
\alias{Abun_CreatBootstrapSample}
\title{R code for create bootstrap sample: only keep not all 0 species and switch column for n1>n2}
\usage{
Abun_CreatBootstrapSample(data, nboots = 0)
}
\arguments{
\item{data}{a matrix/data.frame for species abundance frequency of 2 sites}

\item{nboots}{an integer specifying the number of points of m1}
}
\description{
R code for create bootstrap sample: only keep not all 0 species and switch column for n1>n2
}
