% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.r
\name{Abun_Mix}
\alias{Abun_Mix}
\title{Mixture diversity of 2 communities}
\usage{
Abun_Mix(data, knots = 10, size = NULL)
}
\arguments{
\item{data1}{a matrix/data.frame.}

\item{m}{a matrix specifying the values of m1,m2 where the mixture diversity will be computed.}

\item{n}{a matrix specifying the total observed values : n1,n2.}
}
\value{
a table: \code{$Mix2} a table of mixture diversity for q=0,1,2.
}
\description{
Mixture diversity of 2 communities
}
