% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_subfunction.r
\name{Dq01_ext_inc}
\alias{Dq01_ext_inc}
\title{R code for q=0 2-habitat mixture of  extrapolation curves}
\usage{
Dq01_ext_inc(x1, x2, t_ext, datap)
}
\arguments{
\item{x1}{@param x2 a vector/matrix/list of species abundance frequency}

\item{datap}{probability for x1,x2}

\item{m_ext}{an int matrix for number of individuals for x1,x2(extra parts)}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=0 2-habitat mixture of  extrapolation curves
}
