% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_subfunction.r
\name{Dq01_in}
\alias{Dq01_in}
\title{R code for q=0,1 2-habitat mixture of rarefactioncurves}
\usage{
Dq01_in(x1, x2, mm, q)
}
\arguments{
\item{x1}{a vector/matrix/list of species abundance frequency}

\item{mm}{an int matrix for number of individuals for x1,x2,x3}

\item{q}{a numerical value for Dq}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=0,1 2-habitat mixture of rarefactioncurves
}
\examples{
data(spider)
data1<-spider
}
