% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_subfunction.r
\name{Dq01_in_inc}
\alias{Dq01_in_inc}
\title{R code for q=0,1 2-habitat mixture of rarefactioncurves}
\usage{
Dq01_in_inc(x1, x2, t)
}
\arguments{
\item{x1, x2}{a vector/matrix/list of species incidence frequency and first row is T}

\item{t}{an int matrix for number of individuals for x1,x2}
}
\value{
a vector/matrix/list with species diversity for q=0 and 1
}
\description{
R code for q=0,1 2-habitat mixture of rarefactioncurves
}
\examples{
data(spider)
data1<-spider
}
