% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_subfunction.r
\name{Dq0_ext}
\alias{Dq0_ext}
\title{R code for q=0 2-habitat mixture of  extrapolation curves}
\usage{
Dq0_ext(x1, x2, m_ext, n, datap)
}
\arguments{
\item{x1}{@param x2 a vector/matrix/list of species abundance frequency}

\item{m_ext}{an int matrix for number of individuals for x1,x2(extra parts)}

\item{datap}{probability for x1,x2}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=0 2-habitat mixture of  extrapolation curves
}
