% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_subfunction.r
\name{Dq1_ext}
\alias{Dq1_ext}
\title{R code for q=1 2-habitat mixture of  extrapolation curves}
\usage{
Dq1_ext(x1, x2, m_ext, n, datap)

Dq1_ext(x1, x2, m_ext, n, datap)
}
\arguments{
\item{x1}{@param x2 a vector/matrix/list of species abundance frequency}

\item{m_ext}{an int matrix for number of individuals for x1,x2(extra parts)}

\item{datap}{probability for x1,x2}

\item{x1}{@param x2 a vector/matrix/list of species abundance frequency}

\item{m_ext}{an int matrix for number of individuals for x1,x2(extra parts)}

\item{datap}{probability for x1,x2}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution

a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=1 2-habitat mixture of  extrapolation curves

R code for q=0 Species composition information in a mixed habitat
we can decompose species richness Dq0   of any mixed sample into the sum of three components:
the numbers of shared species, species unique to the intact habitat (Assemblage I), and species unique to the modified habitat (Assemblage II).
}
