% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_subfunction.r
\name{Dq2}
\alias{Dq2}
\title{R code for abundance data q=2 2-habitat mixture of rarefaction and extrapolation curves}
\usage{
Dq2(x1, x2, mm, n1, n2)
}
\arguments{
\item{x1}{@param x2 a vector/matrix/list of species abundance frequency}

\item{mm}{an int matrix for number of individuals for x1,x2}

\item{n1}{@param n2 observations for x1,x2}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for abundance data q=2 2-habitat mixture of rarefaction and extrapolation curves
}
