% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_subfunction.r
\name{Dq2_inci}
\alias{Dq2_inci}
\title{R code for incidence data q=2 2-habitat mixture of rarefaction and extrapolation curves}
\usage{
Dq2_inci(x1, x2, t)
}
\arguments{
\item{x1}{@param x2 a vector/matrix/list of species incidence}

\item{t1}{@param t2 plots for x1,x2}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for incidence data q=2 2-habitat mixture of rarefaction and extrapolation curves
}
