% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{draw.f_Div}
\alias{draw.f_Div}
\title{draw.f_Div_new(data, output, q) plot the outcome of Abundance(abundance data) or Incidence(incidence data).}
\usage{
draw.f_Div(data1, output, type = "abundance", q = 0)
}
\arguments{
\item{output}{matrix form:the outcome of mixture diversity for Abundance/Incidence data by q.}

\item{q}{q=0,1,2}

\item{data}{the Sx2 data used in Abundance/Incidence.}
}
\value{
one plot: for diversity of input q
}
\description{
draw.f_Div_new(data, output, q) plot the outcome of Abundance(abundance data) or Incidence(incidence data).
}
