% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incidence.r
\name{Inci_Mix}
\alias{Inci_Mix}
\title{Mixture diversity of 2 communities}
\usage{
Inci_Mix(data, t, nT)
}
\arguments{
\item{data}{a  Sx2 dataframe, the intact assemblage (main) assemblage should be the first column.}

\item{t}{a matrix specifying the values of t1,t2 where the mixture diversity will be computed.}

\item{nT}{a matrix specifying the values of T1,T2}
}
\value{
a table: \code{$Mix2} a table of mixture diversity for q=0,1,2.
}
\description{
Mixture diversity of 2 communities
}
