% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incidence.r
\name{Incidence}
\alias{Incidence}
\title{Incidence(data, allpts = FALSE, size = NULL, knots = 20 ,nboots = 0) for incidence data, comupute composite diversity of any sample and species composition (shared and unique species)
along with their confidence interval.}
\usage{
Incidence(data, allpts = FALSE, size = NULL, knots = 20,
  nboots = 0, Sdecompose = T)
}
\arguments{
\item{data}{a Sx2 dataframe, the intact assemblage (main) assemblage should be the first column.}

\item{allpts}{specifying whether to compute all combinations of sampling units of two assemblages. Default is FALSE.}

\item{size}{a vector specifying the smapling units of intact (main) assemblage. Default is NULL.}

\item{knots}{the number of points that the mixture diveristy will be computed. Default is 20.}

\item{nboots}{the number of replication bootstrap times. Use 0 to skip bootstrap which might take more time. Default is 0.}
}
\value{
a list containing 4 tables. Th first 3 are diversities of the two assemblages and the mixed one. The 4th table is the species composition of the mixed assemblage. All estimators
are presented along with their confidence interval.
}
\description{
Incidence(data, allpts = FALSE, size = NULL, knots = 20 ,nboots = 0) for incidence data, comupute composite diversity of any sample and species composition (shared and unique species)
along with their confidence interval.
}
