% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{multi.plot}
\alias{multi.plot}
\title{multi.plot(data, ans1, type) plot the outcome of Abundance(abundance data) or Incidence(incidence data).}
\usage{
multi.plot(data, ans1, type = "abundance")
}
\arguments{
\item{data}{the Sx2 data used in Abundance/Incidence.}

\item{ans1}{the outcome of Abundance/Incidence.}

\item{type}{datatype of data, "abundance" or "incidence". Default is "abundance".}
}
\value{
a list containing two plots: the first one for diversity of q = 0, 1, 2 and the second one for species composition.
}
\description{
multi.plot(data, ans1, type) plot the outcome of Abundance(abundance data) or Incidence(incidence data).
}
