% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{plot_comb1}
\alias{plot_comb1}
\title{plot_comb1(data1, output_d, output_p,q) plot species composition.}
\usage{
plot_comb1(data1, output_d, output_p, type = "abundance")
}
\arguments{
\item{output_d}{the outcome of mixture diversity for Abundance/Incidence data by q=0.}

\item{output_p}{the outcome of species composition by q=0.}

\item{data}{the Sx2 data used in Abundance/Incidence.}

\item{q}{q=0: currently only for q=0}
}
\value{
one plot: for diversity of input q
}
\description{
plot_comb1(data1, output_d, output_p,q) plot species composition.
}
