% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_subfunction.r
\name{SDecompose_Inc}
\alias{SDecompose_Inc}
\title{R code for q=0 Species composition information in a mixed habitat}
\usage{
SDecompose_Inc(data, allpts = FALSE, size = NULL, knots = 20,
  q = 0)
}
\arguments{
\item{data}{a matrix/data.frame for species abundance frequency of 2 sites}

\item{size}{an vector}

\item{knots}{an integer specifying the number of points of m1. Default to be 10.}

\item{q}{an integer equal to 0}
}
\description{
R code for q=0 Species composition information in a mixed habitat
}
